+++ RELEASES.rst
@@ -2383,6 +2383,9 @@ Planned
 
 * Add a JSON.stringify() fast path for plain buffers (GH-1238)
 
+* Improve duk_hstring array index handling performance when
+  DUK_USE_HSTRING_ARRIDX is disabled (GH-1274)
+
 * Fix duk_hstring array index check integer overflow, which caused certain
   integer strings (such as '7394299990') to be incorrectly treated as array
   indices (GH-1273, GH-1276)